Add link to deleteAuciton
