+++ baseProject/src/main/java/ch/unibe/ese/team1/model/Ad.java
@@ -6,6 +6,7 @@ import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.ManyToMany;
+import javax.persistence.OneToMany;
 
 import org.hibernate.annotations.Fetch;
 import org.hibernate.annotations.FetchMode;
@@ -18,6 +19,17 @@ public class Ad extends Advertisement{
 	@ManyToMany(cascade = CascadeType.ALL, fetch = FetchType.EAGER)
 	private List<AdPicture> pictures;
 	
+	@OneToMany(mappedBy = "ad", cascade = CascadeType.ALL, fetch = FetchType.EAGER)
+	private List<Visit> visits;
+	
+	public List<Visit> getVisits() {
+		return visits;
+	}
+
+	public void setVisits(List<Visit> visits) {
+		this.visits = visits;
+	}
+
 	public List<AdPicture> getPictures() {
 		return pictures;
 	}
@@ -25,5 +37,7 @@ public class Ad extends Advertisement{
 	public void setPictures(List<AdPicture> pictures) {
 		this.pictures = pictures;
 	}
+	
+	
 	
 }
