+++ baseProject/src/main/java/ch/unibe/ese/team1/model/Advertisement.java
@@ -94,9 +94,6 @@ public abstract class Advertisement {
 	@ManyToOne(optional = false)
 	private User user;
 	
-	@OneToMany(mappedBy = "ad", cascade = CascadeType.ALL, fetch = FetchType.EAGER)
-	private List<Visit> visits;
-	
 	@Column(nullable=false)
 	private boolean auction;
 	
@@ -292,14 +289,6 @@ public abstract class Advertisement {
 			return moveInDate;
 		else
 			return moveOutDate;
-	}
-
-	public List<Visit> getVisits() {
-		return visits;
-	}
-
-	public void setVisits(List<Visit> visits) {
-		this.visits = visits;
 	}
 	
 	public boolean getAuction() {
